requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
babel==2.17.0
    # via
    #   mkdocs-git-revision-date-localized-plugin
    #   mkdocs-material
backrefs==5.9
    # via mkdocs-material
bandit==1.8.6
    # via -r requirements.in
black==25.9.0
    # via -r requirements.in
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.3
    # via requests
click==8.1.7
    # via
    #   black
    #   mkdocs
    #   mkdocs-material
    #   typer
colorama==0.4.6
    # via
    #   bandit
    #   click
    #   mkdocs
    #   mkdocs-material
    #   pytest
    #   radon
    #   tqdm
cryptography==46.0.1
    # via -r requirements.in
distlib==0.4.0
    # via virtualenv
distro==1.9.0
    # via openai
filelock==3.19.1
    # via virtualenv
flake8==7.3.0
    # via -r requirements.in
ghp-import==2.1.0
    # via mkdocs
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via mkdocs-git-revision-date-localized-plugin
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
identify==2.6.14
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via
    #   mkdocs
    #   mkdocs-material
jiter==0.11.0
    # via openai
jsonschema==4.25.1
    # via -r requirements.in
jsonschema-specifications==2025.9.1
    # via jsonschema
libcst==1.8.5
    # via -r requirements.in
mando==0.7.1
    # via radon
markdown==3.9
    # via
    #   mkdocs
    #   mkdocs-material
    #   pymdown-extensions
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   mkdocs
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
mergedeep==1.3.4
    # via
    #   mkdocs
    #   mkdocs-get-deps
mkdocs==1.6.1
    # via
    #   mkdocs-git-revision-date-localized-plugin
    #   mkdocs-material
mkdocs-get-deps==0.2.0
    # via mkdocs
mkdocs-git-revision-date-localized-plugin==1.4.7
    # via -r requirements.in
mkdocs-material==9.6.20
    # via -r requirements.in
mkdocs-material-extensions==1.3.1
    # via mkdocs-material
mypy==1.18.2
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
nodeenv==1.9.1
    # via pre-commit
openai==1.109.1
    # via -r requirements.in
packaging==25.0
    # via
    #   black
    #   mkdocs
    #   pytest
paginate==0.5.7
    # via mkdocs-material
pathspec==0.12.1
    # via
    #   black
    #   mkdocs
    #   mypy
platformdirs==4.4.0
    # via
    #   black
    #   mkdocs-get-deps
    #   virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.0.1
    # via -r requirements.in
pycodestyle==2.14.0
    # via flake8
pycparser==2.23
    # via cffi
pydantic==2.11.9
    # via
    #   -r requirements.in
    #   openai
pydantic-core==2.33.2
    # via pydantic
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via
    #   mkdocs-material
    #   pytest
    #   rich
pymdown-extensions==10.16.1
    # via mkdocs-material
pytest==8.4.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via ghp-import
python-dotenv==1.1.1
    # via -r requirements.in
pytokens==0.1.10
    # via black
pytz==2025.2
    # via mkdocs-git-revision-date-localized-plugin
pyyaml==6.0.3
    # via
    #   bandit
    #   mkdocs
    #   mkdocs-get-deps
    #   pre-commit
    #   pymdown-extensions
    #   pyyaml-env-tag
pyyaml-env-tag==1.1
    # via mkdocs
pyyaml-ft==8.0.0
    # via libcst
radon==6.0.1
    # via -r requirements.in
referencing==0.35.1
    # via
    #   -r requirements.in
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via mkdocs-material
rich==14.1.0
    # via
    #   -r requirements.in
    #   bandit
    #   typer
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
ruff==0.13.1
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   mando
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
stevedore==5.5.0
    # via bandit
tenacity==9.1.2
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typer==0.19.2
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   mypy
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
virtualenv==20.34.0
    # via pre-commit
watchdog==6.0.0
    # via mkdocs
